requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
langchain-community
langchain-cohere
faiss-cpu
cohere
sentence-transformers
scikit-learn
torch
transformers
sqlalchemy==2.0.23
bcrypt==4.1.1
email-validator==2.1.0
python-dateutil==2.8.2